pyproject.toml
[project]
name = "multimodal-ai-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-groq>=0.3.4",
    "langgraph>=0.5.0",
    "python-dotenv>=1.1.1",
]
